+++ src/pages/progress.tsx
@@ -4,11 +4,11 @@ import { Level60Projection } from '@/components/progress/level-60-projection'
 export function Progress() {
   return (
     <div className="space-y-8">
-      {/* Level Timeline */}
-      <LevelTimeline />
-
       {/* Level 60 Projection */}
       <Level60Projection />
+
+      {/* Level Timeline */}
+      <LevelTimeline />
     </div>
   )
 }